+++ service/api/app/internal/logic/transaction/gettxsbypubkeylogic.go
@@ -54,7 +54,7 @@ func (l *GetTxsByPubKeyLogic) GetTxsByPubKey(req *types.ReqGetTxsByPubKey) (*typ
 		}
 		return nil, errorcode.AppErrInternal
 	}
-	txIds, err := l.txDetail.GetTxIdsByAccountIndex(l.ctx, int64(account.AccountIndex))
+	txIds, err := l.txDetail.GetTxIdsByAccountIndex(l.ctx, account.AccountIndex)
 	if err != nil {
 		logx.Errorf("[GetTxDetailByAccountIndex] err: %s", err.Error())
 		if err == errorcode.DbErrNotFound {
